+++ tests/test_code.py
@@ -1,10 +1,15 @@
 from pathlib import Path
 
+import _pytest
+import pytest
+
 HERE = Path(__file__).parent
 TESTDATA_FOLDER = HERE / "testdata" / "folder"
 TESTDATA_SNIPPET = HERE / "testdata" / "snippet"
 
 
+@pytest.mark.pytest_notebook
+@pytest.mark.skipif(_pytest.version_tuple >= (8, 0, 0), reason="Requires pytest version 7")
 def test_monkeypatch_pytest_notebook_treat_cell_exit_as_notebook_skip():
     """
     Verify loading a monkeypatch supporting Jupyter Notebook testing.
@@ -30,6 +35,8 @@ def test_pytest_module_function(request, capsys):
     assert out == "Hallo, Räuber Hotzenplotz.\n"
 
 
+@pytest.mark.pytest_notebook
+@pytest.mark.skipif(_pytest.version_tuple >= (8, 0, 0), reason="Requires pytest version 7")
 def test_pytest_notebook(request):
     """
     Verify executing code cells in an arbitrary Jupyter Notebook.