+++ CHANGES.md
@@ -7,6 +7,7 @@
 - nlp: Updated dependencies langchain, langchain-text-splitters, unstructured
 - CI: Verify compatibility with Python 3.13
 - Testing: Add `pueblo.testing.notebook.{list_notebooks,generate_notebook_tests,run_notebook}`
+- SFA: Added a subsystem for loading single-file applications
 
 ## 2024-03-07 v0.0.9
 - Testing: Add `pueblo.testing.notebook.{list_path,generate_tests}`