+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - ngr: For invoking Elixir, use `mix test --trace`
 - Testing: Unlocked compatibility with pytest version 8
+- SFA: Added a subsystem for loading single-file applications
 
 ## 2024-10-03 v0.0.10
 - nlp: Updated dependencies langchain, langchain-text-splitters, unstructured